Delete analysis_routines.py
Put routines in the runner class, instead of having a separate file
